combat works with ablilities, just needs looting and combat should be done, swapping ablilities appears to work - can't re-arrange inventory currently - need to add a check for statReq, monsters now generate their own weapons and abilities when created, got all the headers ready for doxy comments

+++ src/roomManager.cpp
@@ -75,7 +75,7 @@ void roomManager::enterRoom()
             inventoryManagement(0, 0);
             break;
         case 15:
-          if(input[1] <= chests.size())
+          if(static_cast<unsigned int>(input[1]) <= chests.size())
           {
             chests[input[1] - 1]->open();
             inventoryManagement(1, input[1] - 1);
@@ -86,7 +86,7 @@ void roomManager::enterRoom()
           }
             break;
         case 151:
-          if(input[1] <= chests.size())
+          if(static_cast<unsigned int>(input[1]) <= chests.size())
           {
             chests[input[1] - 1]->open();
             inventoryManagement(1, input[1] - 1);
